+++ src/Pim/Bundle/FilterBundle/Filter/Product/FamilyFilter.php
@@ -6,13 +6,13 @@ use Oro\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
 use Oro\Bundle\FilterBundle\Filter\ChoiceFilter;
 
 /**
- * Product groups filter
+ * Product family filter
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GroupsFilter extends ChoiceFilter
+class FamilyFilter extends ChoiceFilter
 {
     /**
      * {@inheritdoc}
@@ -27,7 +27,7 @@ class GroupsFilter extends ChoiceFilter
         $qb = $ds->getQueryBuilder();
         $ids = $data['value'];
         $repository = $this->util->getProductRepository();
-        $repository->applyFilterByGroupIds($qb, $ids);
+        $repository->applyFilterByFamilyIds($qb, $ids);
 
         return true;
     }